Refactor the list of colors to presentation component.
Eventually I may have another need for a color library of some sort,
but right now I have more usability/ordering concerns for a good user
experience and common options toward the top. That, and the need for
translation of labels, made it seem like this set of constants didn't
belong in the data layer.